requirements.txt
mesa
litellm
matplotlib
numpy
pandas
python-dotenv
markdown
playwright
